completed Hierarchy-Exercise

+++ src/de/auinger/training/java_basics/hierarchy/Vertebrate.java
@@ -2,6 +2,6 @@ package de.auinger.training.java_basics.hierarchy;
 
 public abstract class Vertebrate {
 
-    public abstract int getAverageWeightInKg();
+    public abstract float getAverageWeightInKg();
 
 }
